+++ R/sc_d_recent_history.R
@@ -1,5 +1,6 @@
 recent_history <- function(x, n_recent_obs = 24){
-    result <- x$final$series[,c("y","t","sa")]
+    data_plot <- RJDemetra::get_indicators(x, c("y", "t", "sa"))
+    result <- do.call(ts.union, data_plot)
     if (!is.null(recent_history)) {
         start <- tail(time(result), n_recent_obs)[1]
         result <- ts(tail(result, n_recent_obs), start = start, frequency = frequency(result))