+++ NEWS.md
@@ -1,3 +1,16 @@
+# rjdqa 0.1.6
+
+- new parameter `add_obs_to_forecast` to `simple_dashboard()` and `simple_dashboard2()`.
+
+- new parameter `td_effect` to `simple_dashboard()` and `simple_dashboard2()` to indicate if the residual trading days effect test should be printed.
+By default (`td_effect = NULL`) the test is only printed for monthly series.
+
+- The title of table of outliers of `simple_dashboard2()` now also depend on the number of outliers detected by the model.
+
+- improvement placement of table of outliers in `simple_dashboard2()`.
+
+- improvement of `sc_dashboard()` to handle user-defined calendar regressors.
+
 # rjdqa 0.1.5
 
 - correction of the use of `tail()` on `ts` object to avoid error.